updated pre-sim and calib sim code
I finished creating the baseline catch-at-length code. Key update: after generating proportions harvest- and discard -at-length from the size data, I multiplied these proportions by total *simulated* harvest and discards on *each of the model draws*. Previously, I was pulling only the MRIP point estimates of harvest and discards, resulting in one catch-at-length distribution. With the new approach we are accounting the impact of sampling uncertainty in total harvest and discards on catch-at-length. I think this will also help with convergence in the calibration routine, i.e., ensuring that for each draw, there is enough harvest and discards to reallocate as discards or harvest, respectively, so that the model is able to approximate the MRIP estimate of harvest. I also updated the code for generating catch-per-trip and comparing our simulated harvest, discards, total catch, and directed trips to MRIP point estimates. The comparisons look good; for example, when multiplying a random draw of trips and a random draw of catch-per-trip and doing this 150 times, both the estimate and CIs are very similar to the MRIP estimate and CIs. This gives me confidence that we are appropriately accounting for the sampling uncertainty associated with the MRIP data.