add calendar adjustments, update code
I added to the projection script and data read information about projection year adjustments based on difference in numbers of weekend vs. weekday in the projection year versus the calibration. I also separated the length results processing from the CS/estimated results processing to gain efficiency. I also reverted back to the original way in which I select a number of choice occasions in the projection year to match the number in the calibration year.