Fix length data computation
I fixed a bug in the code for when there was no trips in the calibration year but projected catch in that same period. This caused a mis merge and then missing values in the length dataset and subsequent aggregate calculations. Make sure to change the left_join to right_join on line 337. I also fixed and cleaned up the the aggregating of length data and combining it with CV and predicted trips.