update to projection simulation
key updates. Added code to data_read that removes base outcomes for unnecessary dates. I think this was mucking up the computation of probabilities. Added a few lines to retain baseline trips in addition to predicted trips, as well as a new variable "additional_trips", which is the difference between the two.